Fix behavior of wait for week days and day of month
When using any of the week day method without specifying any time, it will wait for the next
corresponding week day.
When using any of the week day method and specifying a time, it will wait for the next
corresponding week day except if the week day is the same as today and the specified time is later than the current time.
The behavior for the day of month is the same.