Find workflow now throws an exception when API returns an error
Before this change, `Client::findWorkflow()` was returning `null` for
status code that was returned by the API.
Now, when the status code is 404, this method returns `null`.
When the status code is 4xx or 5xx, but not 404, it will throw an
exception.
Otherwise, it will attempt to unserialize the response that was
received.